Moved enums to constants package
