+++ src/main/java/com/softserve/teachua/constants/LogLevel.java
@@ -1,4 +1,4 @@
-package com.softserve.teachua.utils;
+package com.softserve.teachua.constants;
 
 public enum LogLevel {
     ALL, TRACE, DEBUG, INFO, WARN, ERROR, OFF
