+++ src/main/java/com/softserve/teachua/controller/LogLevelController.java
@@ -3,7 +3,7 @@ package com.softserve.teachua.controller;
 import com.softserve.teachua.constants.RoleData;
 import com.softserve.teachua.controller.marker.Api;
 import com.softserve.teachua.service.LogLevelService;
-import com.softserve.teachua.utils.LogLevel;
+import com.softserve.teachua.constants.LogLevel;
 import com.softserve.teachua.utils.annotation.AllowedRoles;
 import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.web.bind.annotation.PatchMapping;
