+++ src/main/java/com/softserve/teachua/service/ExcelParserService.java
@@ -2,7 +2,7 @@ package com.softserve.teachua.service;
 
 import com.softserve.teachua.dto.databaseTransfer.ExcelParsingMistake;
 import com.softserve.teachua.dto.databaseTransfer.ExcelParsingResponse;
-import com.softserve.teachua.utils.excel.ExcelColumn;
+import com.softserve.teachua.constants.excel.ExcelColumn;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.HashMap;
