+++ src/main/java/com/softserve/teachua/service/LogLevelService.java
@@ -1,6 +1,6 @@
 package com.softserve.teachua.service;
 
-import com.softserve.teachua.utils.LogLevel;
+import com.softserve.teachua.constants.LogLevel;
 
 public interface LogLevelService {
     void changeLogLevel(LogLevel logLevel, String packagePath);
