+++ src/main/java/com/softserve/teachua/service/impl/CertificateExcelServiceImpl.java
@@ -12,8 +12,8 @@ import com.softserve.teachua.service.CertificateDataLoaderService;
 import com.softserve.teachua.service.CertificateExcelService;
 import com.softserve.teachua.service.CertificateTemplateService;
 import com.softserve.teachua.service.ExcelParserService;
-import com.softserve.teachua.utils.excel.enums.CertificateExcelColumn;
-import com.softserve.teachua.utils.excel.ExcelColumn;
+import com.softserve.teachua.constants.excel.CertificateExcelColumn;
+import com.softserve.teachua.constants.excel.ExcelColumn;
 import static com.softserve.teachua.utils.validations.CertificateUserNameValidator.NAME_PATTERN;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
