+++ src/main/java/com/softserve/teachua/service/impl/ExcelParserServiceImpl.java
@@ -10,9 +10,9 @@ import com.softserve.teachua.dto.databaseTransfer.model.DistrictExcel;
 import com.softserve.teachua.dto.databaseTransfer.model.LocationExcel;
 import com.softserve.teachua.dto.databaseTransfer.model.StationExcel;
 import com.softserve.teachua.service.ExcelParserService;
-import com.softserve.teachua.utils.ExcelErrorType;
-import com.softserve.teachua.utils.excel.ExcelColumn;
-import com.softserve.teachua.utils.excel.ExcelRowParser;
+import com.softserve.teachua.constants.excel.ExcelErrorType;
+import com.softserve.teachua.constants.excel.ExcelColumn;
+import com.softserve.teachua.utils.ExcelRowParser;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
