+++ src/main/java/com/softserve/teachua/service/impl/LogLevelServiceImpl.java
@@ -4,7 +4,7 @@ import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
 import com.softserve.teachua.service.LogLevelService;
-import com.softserve.teachua.utils.LogLevel;
+import com.softserve.teachua.constants.LogLevel;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
