+++ src/main/java/com/softserve/teachua/service/test/impl/QuestionExcelServiceImpl.java
@@ -9,8 +9,8 @@ import com.softserve.teachua.repository.test.AnswerRepository;
 import com.softserve.teachua.repository.test.QuestionRepository;
 import com.softserve.teachua.service.ExcelParserService;
 import com.softserve.teachua.service.test.QuestionExcelService;
-import com.softserve.teachua.utils.excel.ExcelColumn;
-import com.softserve.teachua.utils.excel.enums.QuestionExcelColumn;
+import com.softserve.teachua.constants.excel.ExcelColumn;
+import com.softserve.teachua.constants.excel.QuestionExcelColumn;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
