+++ src/test/java/com/softserve/teachua/service/CertificateExcelServiceTest.java
@@ -10,8 +10,8 @@ import com.softserve.teachua.dto.certificateExcel.CertificateExcel;
 import com.softserve.teachua.dto.certificateExcel.ExcelParsingResponse;
 import com.softserve.teachua.dto.databaseTransfer.ExcelParsingMistake;
 import com.softserve.teachua.service.impl.CertificateExcelServiceImpl;
-import com.softserve.teachua.utils.excel.ExcelColumn;
-import com.softserve.teachua.utils.excel.enums.CertificateExcelColumn;
+import com.softserve.teachua.constants.excel.ExcelColumn;
+import com.softserve.teachua.constants.excel.CertificateExcelColumn;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import static java.util.Collections.emptyList;
