+++ src/test/java/com/softserve/teachua/service/ExcelParserServiceTest.java
@@ -10,8 +10,8 @@ import static com.softserve.teachua.TestUtils.getIndexes;
 import static com.softserve.teachua.TestUtils.getListOfEmptyCells;
 import com.softserve.teachua.dto.databaseTransfer.ExcelParsingMistake;
 import com.softserve.teachua.service.impl.ExcelParserServiceImpl;
-import com.softserve.teachua.utils.excel.ExcelColumn;
-import com.softserve.teachua.utils.excel.enums.CertificateExcelColumn;
+import com.softserve.teachua.constants.excel.ExcelColumn;
+import com.softserve.teachua.constants.excel.CertificateExcelColumn;
 import java.util.HashMap;
 import java.util.List;
 import org.apache.poi.ss.usermodel.Cell;
